+++ docs/source/changelog.rst
@@ -2,10 +2,12 @@
 Changelog
 =========
 
-v0.5.2
+v0.5.3
 ~~~~~~
 
-- 0561672 Fix:Wrong parameter name
+- 6a26be5 Chg:Wrap the single return value as a list
+- 0b63927 Fix:Item can not extract the data is list type
+- 9deeb5f Chg:Update poetry.lock
 
 .. include:: history.rst
     :start-line: 4